Refactor ConsultancyProjects component: update routing and improve project links for better navigation
